rename single main files to index.ts
